backend/requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
flask-socketio>=5.6.1
pyjwt>=2.9.0
opencv-python>=4.10.0
opencv-contrib-python>=4.10.0
numpy>=1.24.0
pillow>=10.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
scipy>=1.10.0
pydantic>=2.7.0
qrcode[pil]>=7.4.0
ultralytics>=8.2.0
